cbclib_v2 v 0.12.1:
feat: Update dependencies and improve CUDA support

- Updated `pyproject.toml` to include `pkgconfig` and `hdf5plugin` dependencies.
- Enhanced `setup.py` to support CUDA installations and added checks for FFTW library paths.
- Refactored extension definitions to use `CPPExtension` for better C++ and CUDA support.
- Added new tests for CrystFEL geometry file parsing and improved existing tests for data containers and streak finding.
- Updated test cases to reflect changes in the API and ensure compatibility with new features.

+++ cbclib_v2/_src/src/signal_proc.cpp
@@ -20,7 +20,7 @@ py::array_t<T> binterpolate(py::array_t<T> inp, std::vector<py::array_t<U>> grid
 
     auto npts = coords.size() / ndim;
     if (ndim > 1) check_dimension("coords", coords.ndim() - 1, coords.shape(), ndim);
-    else if (*(coords.shape() + coords.ndim() - 1) != ndim)
+    else if (static_cast<size_t>(*(coords.shape() + coords.ndim() - 1)) != ndim)
     {
         std::vector<size_t> shape {coords.shape(), coords.shape() + coords.ndim()};
         shape.push_back(ndim);
